WIP: use Long instead of BigDecimal

+++ src/main/kotlin/com/ampnet/crowdfundingbackend/service/BlockchainService.kt
@@ -1,7 +1,5 @@
 package com.ampnet.crowdfundingbackend.service
 
-import java.math.BigDecimal
-
 interface BlockchainService {
-    fun getBalance(address: String): BigDecimal
+    fun getBalance(address: String): Long
 }
